SHOP COMPONENT Moved shop collections data to Redux
