+++ src/redux/root-reducer.js
@@ -6,6 +6,7 @@ import storage from "redux-persist/lib/storage";
 import userReducer from "./user/user.reducer";
 import cartReducer from "./cart/cart.reducer";
 import directoryReducer from "./directory/directory.reducer";
+import shopReducer from "./shop/shop.reducer";
 
 const persistConfig = {
 	key: "root",
@@ -16,6 +17,7 @@ const persistConfig = {
 const rootReducer = combineReducers({
 	user: userReducer,
 	cart: cartReducer,
-	directory: directoryReducer
+	directory: directoryReducer,
+	shop: shopReducer
 });
 export default persistReducer(persistConfig, rootReducer);
