+++ src/redux/shop/shop.selector.js
@@ -1,0 +1,8 @@
+import { createSelector } from "reselect";
+
+const selectShop = state => state.shop;
+
+export const selectCollectionsItems = createSelector(
+	[selectShop],
+	shop => shop.collections
+);
